+++ src/components/editTestCase/qdm/LeftPanel/ElementsTab/Demographics/DemographicsSection.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import ElementSection from "../ElementSection";
+import ElementSection from "../../../../../../util/ElementSection";
 import { Select, InputLabel } from "@madie/madie-design-system/dist/react";
 import FormControl from "@mui/material/FormControl";
 import InputAdornment from "@mui/material/InputAdornment";
