--- src/components/editTestCase/qiCore/LeftPanel/ElementsTab/ElementSection.scss
@@ -1,34 +1,0 @@
-$blue-60: #0073c8;
-$blue-70: #125496;
-$grey-19: #dddddd;
-
-.test-case-tab-heading {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 48px;
-  > .heading-row {
-    display: flex;
-    border-bottom: solid 1px $grey-19;
-    flex-direction: row;
-    flex-grow: 1;
-    justify-content: space-between;
-    align-items: center;
-    cursor: pointer;
-    > .header {
-      color: $blue-70;
-    }
-    > .chevron-display {
-      color: $blue-60;
-      transition: all 0.2s ease;
-      &.open {
-        transform: rotate(90deg);
-      }
-    }
-  }
-  > .growing-div {
-    flex-direction: column;
-    &.open {
-      margin-top: 35px;
-    }
-  }
-}
