+++ src/components/editTestCase/qiCore/LeftPanel/ElementsTab/Elements/ElementsSection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ElementSection from "../ElementSection";
+import ElementSection from "../../../../../../util/ElementSection";
 
 const ElementsSection = ({ canEdit }) => {
   return (
